# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Ecash wallet and mint with Bitcoin Lightning support.",
        "title": "Cashu (Gongo) golang mint",
        "contact": {
            "url": "https://8333.space:3338"
        },
        "version": "0.0.1"
    },
    "paths": {
        "/check": {
            "post": {
                "description": "Get currently available public keys",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "POST"
                ],
                "summary": "Check spendable",
                "parameters": [
                    {
                        "description": "Model containing proofs to check",
                        "name": "CheckRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/mint.CheckRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/mint.CheckResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/mint.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/checkfees": {
            "post": {
                "description": "Check fees for lightning payment.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "POST"
                ],
                "summary": "Check Fee",
                "parameters": [
                    {
                        "description": "Model containing lightning invoice",
                        "name": "CheckFeesRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/mint.CheckFeesRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/mint.CheckFeesResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/mint.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/keys": {
            "get": {
                "description": "Get the public keys of the mint",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "GET"
                ],
                "summary": "Keys",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/mint.GetKeysResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/mint.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/melt": {
            "post": {
                "description": "Requests tokens to be destroyed and sent out via Lightning.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "POST"
                ],
                "summary": "Melt",
                "parameters": [
                    {
                        "description": "Model containing proofs to melt",
                        "name": "MeltRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/mint.MeltRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/mint.MeltResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/mint.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/mint": {
            "get": {
                "description": "Requests the minting of tokens belonging to a paid payment request.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "GET"
                ],
                "summary": "Request Mint",
                "parameters": [
                    {
                        "type": "string",
                        "description": "amount of the mint",
                        "name": "amount",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/mint.GetMintResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/mint.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Requests the minting of tokens belonging to a paid payment request.\n\nParameters: pr: payment_request of the Lightning paid invoice.\n\nBody (JSON): payloads: contains a list of blinded messages waiting to be signed.\n\nNOTE:\n\n* This needs to be replaced by the preimage otherwise someone knowing the payment_request can request the tokens instead of the rightful owner.\n* The blinded message should ideally be provided to the server before payment in the GET /mint endpoint so that the server knows to sign only these tokens when the invoice is paid.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "POST"
                ],
                "summary": "Mint",
                "parameters": [
                    {
                        "description": "Model containing proofs to mint",
                        "name": "core.BlindedMessages",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/core.BlindedMessage"
                            }
                        }
                    },
                    {
                        "type": "string",
                        "description": "payment hash for the mint",
                        "name": "payment_hash",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/core.BlindedMessage"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/mint.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/split": {
            "post": {
                "description": "Requetst a set of tokens with amount \"total\" to be split into two newly minted sets with amount \"split\" and \"total-split\".",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "POST"
                ],
                "summary": "Split your proofs",
                "parameters": [
                    {
                        "description": "Model containing proofs to split",
                        "name": "SplitRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/mint.SplitRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/mint.SplitResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/mint.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "core.BlindedMessage": {
            "type": "object",
            "properties": {
                "B_": {
                    "type": "string"
                },
                "amount": {
                    "type": "integer"
                }
            }
        },
        "core.Proof": {
            "type": "object",
            "properties": {
                "C": {
                    "type": "string"
                },
                "amount": {
                    "type": "integer"
                },
                "secret": {
                    "type": "string"
                }
            }
        },
        "mint.CheckFeesRequest": {
            "type": "object",
            "properties": {
                "pr": {
                    "type": "string"
                }
            }
        },
        "mint.CheckFeesResponse": {
            "type": "object",
            "properties": {
                "fee": {
                    "type": "integer"
                }
            }
        },
        "mint.CheckRequest": {
            "type": "object",
            "properties": {
                "proofs": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/core.Proof"
                    }
                }
            }
        },
        "mint.CheckResponse": {
            "type": "object",
            "additionalProperties": {
                "type": "boolean"
            }
        },
        "mint.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "error": {
                    "type": "string"
                }
            }
        },
        "mint.GetKeysResponse": {
            "type": "object",
            "additionalProperties": {
                "type": "string"
            }
        },
        "mint.GetMintResponse": {
            "type": "object",
            "properties": {
                "hash": {
                    "type": "string"
                },
                "pr": {
                    "type": "string"
                }
            }
        },
        "mint.MeltRequest": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "integer"
                },
                "invoice": {
                    "type": "string"
                },
                "proofs": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/core.Proof"
                    }
                }
            }
        },
        "mint.MeltResponse": {
            "type": "object",
            "properties": {
                "paid": {
                    "type": "boolean"
                },
                "preimage": {
                    "type": "string"
                }
            }
        },
        "mint.SplitRequest": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "integer"
                },
                "outputs": {
                    "type": "object",
                    "properties": {
                        "blinded_messages": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/core.BlindedMessage"
                            }
                        }
                    }
                },
                "proofs": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/core.Proof"
                    }
                }
            }
        },
        "mint.SplitResponse": {
            "type": "object",
            "properties": {
                "fst": {
                    "type": "string"
                },
                "snd": {
                    "type": "string"
                }
            }
        }
    }
}